change git action yaml

+++ src/include/homestore/btree/detail/btree_internal.hpp
@@ -246,8 +246,8 @@ struct BtreeConfig {
     uint32_t m_max_merge_nodes{3};
 #ifdef _PRERELEASE
     // These are for testing purpose only
-    uint64_t m_max_keys_in_node{0};
-    uint64_t m_min_keys_in_node{0};
+    uint32_t m_max_keys_in_node{0};
+    uint32_t m_min_keys_in_node{0};
 #endif
     bool m_rebalance_turned_on{false};
